services/ai_processor_service/requirements.txt
pika==1.3.2
pydantic==2.5.3
google-generativeai==0.8.3